# Description

At Geofabrik, we create raster tiles from vector tiles using Mapnik and its OGR input plugin which cannot sort the features. However, some features should be sorted. Cities by population (e.g. Frankfurt am Main vs. Offenbach am Main), areas by their size (lacking usage of multipolygons by mappers) are common examples.

This pull request adds a callback function to store a number (32 bit unsigned integer) which can be calculated by the Lua script. I decided to use 32 bit only because it is sufficient for most use cases and therefore a good compromise between memory and perfection. For areas, it' wise to convert the sizes from square metre to hectar to get the number smaller by the factor of 10^4. For cities.

# Performance Tests

Performance tests with 64 cores, 504 GB RAM, cache on NVMe (if enabled), configuration from [geofabrik-basicvt-tilemaker commit fbbe9a66d30cacb3314a6d43d5e5a37e95254b8e](https://github.com/geofabrik/geofabrik-basicvt-tilemaker/tree/fbbe9a66d30cacb3314a6d43d5e5a37e95254b8e):

`tilemaker --input /planet/current-planet.osm.pbf --output /mnt/vectortiles/gf-vt-basic-planet/ --store /nvme/michael/tilemaker-cache.dat --config config.json --process process.lua --bbox -180,-86,180,86`

* this branch: 15 hours 45 minutes, 328 GB RAM
* master branch (calls to `SortableNumber` commented out): 16 hours 18 seconds, 326 GB RAM

Performance tests without `--store` for the whole planet are still running.

`tilemaker --input /extracts/europe/germany.osm.pbf --output /mnt/vectortiles/gf-vt-basic-germany/ --store /nvme/michael/tilemaker-cache.dat --config config.json --process process.lua`

* this branch: 14 minutes 8 seconds, 27.5 GB RAM
* master branch (calls to `SortableNumber` commented out): 14 minutes 8 seconds, 27.2 GB RAM

`tilemaker --input /extracts/europe/germany.osm.pbf --output /mnt/vectortiles/gf-vt-basic-germany --config config.json --process process.lua`:

* this branch: 14 minutes 2 seconds, 36 GB RAM
* master branch (calls to `SortableNumber` commented out): 13 minutes 57 seconds, 36 GB RAM