@kleunen 2021-09-28
Are you sure you want to store uint32? And not a float? Because float supports a much wider range of numbers. If you calculate the area, you may not know how big this number might become.