@systemed 2021-10-04
Thanks for this - looks like a good idea.

Could we achieve the same just by using the existing `ZOrder`, perhaps expanded to a uint32 or float? Ultimately what you're doing here, I think, is providing a way to sort objects which works pretty much the same way as ZOrder:

	if (x->z_order < y->z_order) return true;
 	if (x->z_order > y->z_order) return false;
 	// Comparison inverted for area because large features should appear first in the vector tile
 	// if the area was set.
 	if (x->sortable_number > y->sortable_number) return true;
 	if (x->sortable_number < y->sortable_number) return false;

For place nodes you probably wouldn't need any other ZOrder anyway. In the event that you had something that you wanted to sort both by ZOrder and the sortable property, you'd just do something like (ZOrder*100-population/500) or whatever.